after incomplete, no longer restarts for today
Batch checks to see if there is an incomplete scan in the last 10 days. If there is, it will finish it first. Afterwards, it used to start a 'fresh run' for today. This is now disabled, the batch either finishes an old incomplete run, OR it does a fresh run if no incomplete run was found.